+++ src/main/webapp/ui/src/eln/gallery/components/useIrods.js
@@ -2,13 +2,16 @@
 
 import axios, { type Axios } from "axios";
 import React from "react";
-import Result from "../../util/result";
-import * as Parsers from "../../util/parsers";
-import * as FetchingData from "../../util/fetchingData";
-import useOauthToken from "../../common/useOauthToken";
-import { Optional } from "../../util/optional";
-import AlertContext, { mkAlert, type Alert } from "../../stores/contexts/Alert";
-import * as ArrayUtils from "../../util/ArrayUtils";
+import Result from "../../../util/result";
+import * as Parsers from "../../../util/parsers";
+import * as FetchingData from "../../../util/fetchingData";
+import useOauthToken from "../../../common/useOauthToken";
+import { Optional } from "../../../util/optional";
+import AlertContext, {
+  mkAlert,
+  type Alert,
+} from "../../../stores/contexts/Alert";
+import * as ArrayUtils from "../../../util/ArrayUtils";
 
 type Link = {| operation: string, href: string |};
